fyp_multimodal_model/requirements.txt
pandas
numpy
scikit-learn
lightgbm
shap
gensim
tqdm
torch
torchvision
Pillow
joblib
selenium
beautifulsoup4
webdriver-manager
flask
